♻️ Add helpers for parameterValidator
